Changed primary keys to SERIAL and included flyway access token

+++ src/database/migrations/V4__create_customers_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "Customers" (
-  "CustomerID" bigint PRIMARY KEY,
+  "CustomerID" SERIAL PRIMARY KEY,
   "Name" varchar(255),
   "Email" varchar(255) UNIQUE,
   "PhoneNumber" varchar(255)
