+++ src/database/migrations/V7__create_orders_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "Orders" (
-  "OrderID" bigint PRIMARY KEY NOT NULL,
+  "OrderID" SERIAL PRIMARY KEY NOT NULL,
   "CustomerID" bigint NOT NULL,
   "OrderDate" date NOT NULL,
   "TotalAmount" decimal(8,2) NOT NULL,
